requirements.txt
streamlit
mammoth
beautifulsoup4
docxtpl
python-docx